+++ src/main/java/CountingDaoFactory.java
@@ -5,7 +5,7 @@ import user.connection.ConnectionMaker;
 import user.connection.CountingConnectionMaker;
 import user.connection.SimpleConnectionMaker;
 import user.dao.JdbcContext;
-import user.dao.UserDao;
+import user.dao.UserDaoJdbc;
 
 import javax.sql.DataSource;
 import java.sql.Driver;
@@ -14,8 +14,8 @@ import java.sql.SQLException;
 @Configuration
 public class CountingDaoFactory {
     @Bean
-    public UserDao userDao() throws SQLException {
-        return new UserDao(dataSource());
+    public UserDaoJdbc userDao() throws SQLException {
+        return new UserDaoJdbc(dataSource());
     }
 
     @Bean
