+++ src/main/java/DaoFactory.java
@@ -4,7 +4,7 @@ import org.springframework.jdbc.datasource.SimpleDriverDataSource;
 import user.connection.ConnectionMaker;
 import user.connection.SimpleConnectionMaker;
 import user.dao.JdbcContext;
-import user.dao.UserDao;
+import user.dao.UserDaoJdbc;
 
 import javax.sql.DataSource;
 import java.sql.Driver;
@@ -13,8 +13,8 @@ import java.sql.SQLException;
 @Configuration
 public class DaoFactory {
     @Bean
-    public UserDao userDao() throws SQLException {
-        return new UserDao(dataSource());
+    public UserDaoJdbc userDao() throws SQLException {
+        return new UserDaoJdbc(dataSource());
     }
 
     @Bean
